+++ src/Integration.Vsix.UnitTests/Integration.Vsix.UnitTests.csproj
@@ -1,7 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <Import Project="..\SonarLint.Test.props" />
-  <!--<Import Project="..\SonarLint.VsPkgRefs.props" />-->
 
   <PropertyGroup Label="Specify which VS-specific refs are required">
     <RequiresVCProjectEngine>true</RequiresVCProjectEngine>
@@ -13,7 +12,7 @@
     <ProjectGuid>{203F7669-02B6-4314-BD11-20E8D7C19FBC}</ProjectGuid>
     <RootNamespace>SonarLint.VisualStudio.Integration.UnitTests</RootNamespace>
     <AssemblyName>SonarLint.VisualStudio.Integration.Vsix.UnitTests</AssemblyName>
-        <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
+    <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
     <GenerateBindingRedirectsOutputType>true</GenerateBindingRedirectsOutputType>
   </PropertyGroup>
   
@@ -47,19 +46,16 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
-    <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
-    <PackageReference Include="MSTest.TestAdapter" Version="1.4.0" />
-    <PackageReference Include="MSTest.TestFramework" Version="1.4.0" />
     <PackageReference Include="System.Collections.Immutable" Version="1.1.37" />
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="1.3.2" />
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="1.3.2" />
     <PackageReference Include="Microsoft.CodeAnalysis.Workspaces.Common" Version="1.3.2" />
     <PackageReference Include="Microsoft.Composition" Version="1.0.30" />
     <PackageReference Include="NuGet.Core" Version="2.12.0" />
+    <PackageReference Include="NuGet.VisualStudio" Version="3.3.0" />
     <PackageReference Include="System.IO.Abstractions" Version="9.0.4" />
     <PackageReference Include="System.IO.Abstractions.TestingHelpers" Version="9.0.4" />
     <PackageReference Include="System.Reflection.Metadata" Version="1.2.0" />
-    <PackageReference Include="NuGet.VisualStudio" Version="3.3.0" />
   </ItemGroup>
 
   <!-- Only need to reference the analyzer common assembly for testing. -->
@@ -80,6 +76,10 @@
 
   <ItemGroup Label="VSSDK Test framework">
     <PackageReference Condition="$(VsTargetVersion) != '2022'" Include="Microsoft.VisualStudio.Sdk.TestFramework" Version="16.5.22-beta" />
+
+    <!-- NOTE: the "17.0.8.alpha" package is not available from Microsoft in the public NuGet
+         feed. It was created by building and packaging the MS source code at the following
+         commit: https://github.com/microsoft/vssdktestfx/commit/bb86ac65c2195e81e8131865acbbc51f448d9ce8-->
     <PackageReference Condition="$(VsTargetVersion) == '2022'" Include="Microsoft.VisualStudio.Sdk.TestFramework" Version="17.0.8-alpha" />
   </ItemGroup>
